Added requirements to make a executable
